# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Consolidated User Service API
  version: 1.0.0
  description: |
    API for user authentication, profile management, and session handling.
    
    # User Types and Flows
    
    ## Regular User Flow
    1. Sign up with email/phone
    2. Verify email/phone
    3. Create basic profile
    4. Login and access user features
    
    ## Creator Flow
    1. Sign up with email/phone
    2. Verify email/phone
    3. Create creator profile with services
    4. Set location and availability
    5. Login and access creator features
    
    # API Categories
    
    ## Authentication
    - Email/phone signup and verification
    - Profile creation
    - Login (email/password, phone/OTP)
      - Password reset
      - Google OAuth2
    
    ## User Features
    - View and update profile
    - Manage preferences
    - View creator profiles
    - Book services
    
    ## Creator Features
    - Manage services offered
    - Set location and availability
    - View bookings
    - Manage profile visibility
    
    ## Admin Features
    - User management
    - Content moderation
    - Analytics and reporting
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email with code
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
              required: [email, code]
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /auth/signin/email:
    post:
      tags:
        - Authentication
      summary: Sign in with email and password
      operationId: signinEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
              required: [email, password]
      responses:
        '200':
          description: JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/signin/phone:
    post:
      tags:
        - Authentication
      summary: Sign in with phone (send OTP)
      operationId: signinPhone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: Phone number in E.164 format
              required: [phone]
      responses:
        '200':
          description: OTP sent to phone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Start password reset (send code)
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: Reset code sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/verify-reset-code:
    post:
      tags:
        - Authentication
      summary: Verify password reset code
      operationId: verifyResetCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
              required: [email, code]
      responses:
        '200':
          description: Reset code verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                new_password:
                  type: string
                  minLength: 8
              required: [email, code, new_password]
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request - The server cannot process the request due to a client error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication is required and has failed or has not been provided
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Internal Server Error - The server has encountered a situation it does not know how to handle
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message indicating the result of the operation
      required:
        - message
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for authentication
        expires_at:
          type: string
          format: date-time
          description: Token expiration time
      required:
        - token
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details if available
      required:
        - error
        - message
...